pyproject.toml
[tool.poetry]
name = "绘梨花QQBot|ErikaBot"
version = "1.1.0"
description = "A tencent QQ robot"
authors = ["SlieFamily <sliewdyinwhite@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nb-cli = "^0.6.9"
selenium = "^4.1.3"
nonebot-adapter-onebot = "^2.2.3"
opencv-python = "^4.5.5"
nonebot-plugin-guild-patch = "^0.1.3"
nonebot-plugin-apscheduler = "^0.3.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
plugins = ["nonebot_plugin_guild_patch"]
plugin_dirs = ["src/plugins"]

[nonebot.plugins]
plugin_dirs = ["src/plugins"]
